refactor: quebra LigaDetailPage, ConfrontoSection e AdminPage em arquivos menores
Sem mudança de comportamento, props ou estilo — apenas extração estrutural e
ajuste de imports. Build (tsc + vite) e lint verdes (mesma contagem de avisos
pré-existente).

- leagues: extrai ClassificacaoTab, MembrosTab, CompeticoesTab e EscudoStudio
  para src/features/leagues/tabs/; LigaDetailPage fica só como página +
  composição (973 -> 398 linhas).
- confronto: extrai SorteioPanel, ScheduledView e DrawnView para arquivos
  próprios; ConfrontoSection vira só o dispatcher (916 -> 69). A matemática
  pura de viabilidade/rodadas vai para helpers testáveis em build.ts
  (confrontoViability, testViability) — sem alterar resultados.
- admin: extrai LigasAdmin, CompeticoesAdmin e UsuariosAdmin; AdminPage fica só
  como shell + abas (592 -> 44). PaymentAdmin (Pgto/Descontos) já era separado.
