+++ src/features/leagues/LigaDetailPage.tsx
@@ -1,54 +1,32 @@
 import { useEffect, useMemo, useState } from "react";
 import { useParams, useNavigate, useSearchParams } from "react-router-dom";
 import { useQueryClient } from "@tanstack/react-query";
-import {
-  ArrowLeft,
-  Copy,
-  Clock,
-  Check,
-  Trash2,
-  ShieldCheck,
-  Plus,
-  LogOut,
-  Palette,
-  Sparkles,
-} from "lucide-react";
+import { ArrowLeft, Copy, Clock, LogOut, Palette, Sparkles } from "lucide-react";
 import { Page } from "@/components/layout/Page";
 import { Card } from "@/components/ui/Card";
 import { Button } from "@/components/ui/Button";
-import { Badge } from "@/components/ui/Badge";
-import { Avatar } from "@/components/ui/Avatar";
 import { Skeleton } from "@/components/ui/Skeleton";
 import { EmptyState } from "@/components/ui/EmptyState";
 import { SegmentedControl } from "@/components/ui/SegmentedControl";
 import { ConfirmDialog } from "@/components/ui/ConfirmDialog";
 import { Escudo } from "@/components/ui/Escudo";
-import { CrestEditor } from "@/components/ui/CrestEditor";
-import { FLAMULA_SHAPES } from "@/lib/crest";
 import { useToast } from "@/components/ui/Toast";
 import { useAuth } from "@/features/auth/AuthProvider";
-import { useCompetitions, findWorldCupCompetition } from "@/features/matches/api";
-import { useStandings } from "./api";
-import { StandingsTable } from "@/features/standings/StandingsTable";
-import { ConfrontoSection } from "@/features/confronto/ConfrontoSection";
 import {
+  useStandings,
   useLeague,
   useLeagueMembers,
   useLeagueCompetitions,
-  useUpdateMember,
-  useRemoveMember,
   useLeaveLeague,
   useSetConfrontoEnabled,
-  useAddLeagueCompetition,
-  useDeleteLeagueCompetition,
-  useUpdateLeagueLogo,
   useLeagueCheckout,
 } from "./api";
 import { usePaymentSettings, useSimulatePayment, useCompLeague } from "@/features/payments/api";
-import { useNamePrefixes, requiredPrefix, NAME_PREFIX_DEFAULTS } from "./naming";
 import { RefundFederationButton } from "./RefundFederationButton";
-import type { LeagueMode } from "@/lib/types";
-import { cn } from "@/lib/utils";
+import { ClassificacaoTab } from "./tabs/ClassificacaoTab";
+import { MembrosTab } from "./tabs/MembrosTab";
+import { CompeticoesTab } from "./tabs/CompeticoesTab";
+import { EscudoStudio } from "./tabs/EscudoStudio";
 
 type Tab = "classificacao" | "membros" | "competicoes";
 
@@ -416,558 +394,5 @@ export function LigaDetailPage() {
         onCancel={() => setLeaveOpen(false)}
       />
     </Page>
-  );
-}
-
-function ClassificacaoTab({
-  comps,
-  activeLcId,
-  onSelect,
-  standings,
-  loading,
-  currentUserId,
-  isAdmin,
-  leagueId,
-  memberCount,
-}: {
-  comps: {
-    id: string;
-    name: string;
-    mode: string;
-    competition_id: string;
-    confronto_state?: string;
-    participant_mode?: string;
-    liga_format?: string;
-    scheduled_draw_at?: string | null;
-  }[];
-  activeLcId?: string;
-  onSelect: (id: string) => void;
-  standings: ReturnType<typeof useStandings>["data"];
-  loading: boolean;
-  currentUserId?: string;
-  isAdmin: boolean;
-  leagueId: string;
-  memberCount: number;
-}) {
-  // Separa as disputas por tipo: Pontos (corrida por campeonato) e Confronto (Liga/Copa).
-  const isConfronto = (m: string) => m === "liga" || m === "cup";
-  const pontos = comps.filter((c) => !isConfronto(c.mode));
-  // Rascunho (em configuração) só aparece pro admin; pros demais, só quando liberado.
-  const confronto = comps.filter(
-    (c) => isConfronto(c.mode) && (isAdmin || c.confronto_state !== "draft"),
-  );
-  const hasBoth = pontos.length > 0 && confronto.length > 0;
-  const activeComp = comps.find((c) => c.id === activeLcId);
-  const [view, setView] = useState<"pontos" | "confronto">(
-    activeComp && isConfronto(activeComp.mode)
-      ? "confronto"
-      : pontos.length === 0 && confronto.length > 0
-        ? "confronto"
-        : "pontos",
-  );
-
-  const group: "pontos" | "confronto" = hasBoth
-    ? view
-    : confronto.length > 0
-      ? "confronto"
-      : "pontos";
-  const list = group === "confronto" ? confronto : pontos;
-  const active = list.find((c) => c.id === activeLcId) ?? list[0];
-
-  // Mantém a seleção do pai (e a query de classificação) em sincronia com a aba visível.
-  useEffect(() => {
-    if (active && active.id !== activeLcId) onSelect(active.id);
-  }, [active, activeLcId, onSelect]);
-
-  if (comps.length === 0) {
-    return (
-      <EmptyState
-        title="Sem competições"
-        description="Um administrador precisa vincular uma competição a esta federação."
-      />
-    );
-  }
-
-  return (
-    <div className="space-y-3">
-      {hasBoth && (
-        <>
-          <SegmentedControl<"pontos" | "confronto">
-            value={view}
-            onChange={setView}
-            options={[
-              { value: "pontos", label: "Pontos" },
-              { value: "confronto", label: "Confronto" },
-            ]}
-          />
-          <p className="px-1 text-xs leading-relaxed text-ink-500">
-            {group === "pontos"
-              ? "Corrida de pontos por campeonato — quem somou mais lidera."
-              : "Duelos diretos: Liga (tabela 3/1/0) e Copa (mata-mata)."}
-          </p>
-        </>
-      )}
-
-      {list.length > 1 && (
-        <div className="no-scrollbar -mx-4 flex gap-2 overflow-x-auto px-4">
-          {list.map((c) => (
-            <button
-              key={c.id}
-              onClick={() => onSelect(c.id)}
-              className={cn(
-                "inline-flex shrink-0 items-center gap-1.5 rounded-pill border px-3 py-1.5 text-sm font-semibold transition",
-                active?.id === c.id
-                  ? "border-brand-600 bg-brand-50 text-brand-700"
-                  : "border-ink-200 bg-surface text-ink-600",
-              )}
-            >
-              {c.name}
-              {isConfronto(c.mode) && (
-                <span className="rounded-pill bg-ink-100 px-1.5 text-[10px] font-bold uppercase tracking-wide text-ink-500">
-                  {c.mode === "cup" ? "Copa" : "Liga"}
-                </span>
-              )}
-            </button>
-          ))}
-        </div>
-      )}
-
-      {active && isConfronto(active.mode) ? (
-        <ConfrontoSection
-          lcId={active.id}
-          leagueId={leagueId}
-          competitionId={active.competition_id}
-          mode={active.mode}
-          state={active.confronto_state ?? "draft"}
-          memberCount={memberCount}
-          isAdmin={isAdmin}
-          currentUserId={currentUserId}
-          participantMode={active.participant_mode ?? "admin"}
-          ligaFormat={active.liga_format ?? "partial"}
-          scheduledDrawAt={active.scheduled_draw_at ?? null}
-        />
-      ) : loading ? (
-        <Skeleton className="h-64 w-full" />
-      ) : standings && standings.length > 0 ? (
-        <StandingsTable rows={standings} currentUserId={currentUserId} />
-      ) : (
-        <EmptyState title="Sem pontos ainda" description="A classificação aparece após os jogos." />
-      )}
-    </div>
-  );
-}
-
-function MembrosTab({
-  members,
-  isAdmin,
-}: {
-  members: ReturnType<typeof useLeagueMembers>["data"] & object;
-  isAdmin: boolean;
-}) {
-  const update = useUpdateMember();
-  const remove = useRemoveMember();
-  const { toast } = useToast();
-  const list = members ?? [];
-
-  return (
-    <ul className="space-y-2">
-      {list.map((m) => (
-        <Card key={m.id} className="flex items-center gap-3 p-3">
-          <Avatar src={m.profile?.avatar_url} name={m.profile?.display_name} size="md" />
-          <div className="min-w-0 flex-1">
-            <p className="truncate font-semibold text-ink-900">{m.profile?.display_name}</p>
-            <div className="flex items-center gap-1.5">
-              {m.role !== "member" && (
-                <Badge tone="brand">{m.role === "owner" ? "Dono" : "Admin"}</Badge>
-              )}
-              {m.status === "pending" && <Badge tone="gold">pendente</Badge>}
-            </div>
-          </div>
-          {isAdmin && m.role !== "owner" && (
-            <div className="flex gap-1">
-              {m.status === "pending" && (
-                <Button
-                  size="icon"
-                  variant="ghost"
-                  aria-label="Aprovar"
-                  onClick={() =>
-                    update.mutate(
-                      { memberId: m.id, status: "active" },
-                      { onSuccess: () => toast("Membro aprovado!", "success") },
-                    )
-                  }
-                >
-                  <Check className="size-4 text-grass-600" />
-                </Button>
-              )}
-              {m.status === "active" && (
-                <Button
-                  size="icon"
-                  variant="ghost"
-                  aria-label={m.role === "admin" ? "Rebaixar" : "Promover a admin"}
-                  onClick={() =>
-                    update.mutate({
-                      memberId: m.id,
-                      role: m.role === "admin" ? "member" : "admin",
-                    })
-                  }
-                >
-                  <ShieldCheck
-                    className={cn("size-4", m.role === "admin" ? "text-brand-600" : "text-ink-300")}
-                  />
-                </Button>
-              )}
-              <Button
-                size="icon"
-                variant="ghost"
-                aria-label="Remover"
-                onClick={() =>
-                  remove.mutate(m.id, { onSuccess: () => toast("Membro removido.", "info") })
-                }
-              >
-                <Trash2 className="size-4 text-flame-500" />
-              </Button>
-            </div>
-          )}
-        </Card>
-      ))}
-    </ul>
-  );
-}
-
-// Limite inicial: 1 competição por federação. Quando a base de usuários crescer e
-// os modos extras (Liga / Copa) estiverem prontos, soltamos o limite e habilitamos
-// os outros modos de disputa.
-const MAX_COMPETITIONS_PER_LEAGUE = 1;
-
-function CompeticoesTab({
-  leagueId,
-  comps,
-  confrontoEnabled,
-}: {
-  leagueId: string;
-  comps: { id: string; name: string; mode: string }[];
-  confrontoEnabled: boolean;
-}) {
-  const { data: competitions } = useCompetitions();
-  const add = useAddLeagueCompetition();
-  const del = useDeleteLeagueCompetition();
-  const { toast } = useToast();
-  const [open, setOpen] = useState(false);
-  const [competitionId, setCompetitionId] = useState("");
-  const [name, setName] = useState("");
-  const [toDelete, setToDelete] = useState<{ id: string; name: string } | null>(null);
-
-  // Federação comum: modo fixo em "points" (Liga/Copa ficam como "em breve").
-  // Federação habilitada pelo admin (confronto_enabled): o admin escolhe Pontos ou
-  // Confronto (Liga/Copa) e pode criar várias competições (sem o limite do MVP).
-  const [tipo, setTipo] = useState<"pontos" | "confronto">("pontos");
-  const [formato, setFormato] = useState<"liga" | "cup">("liga");
-  const [participantMode, setParticipantMode] = useState<"admin" | "optin">("admin");
-  const mode: LeagueMode = !confrontoEnabled || tipo === "pontos" ? "points" : formato;
-  const isConfrontoNew = mode === "liga" || mode === "cup";
-
-  // Prefixo do tipo (Bolão/Liga/Copa) entra como badge fixa — a pessoa digita só o resto.
-  const { data: prefixesData } = useNamePrefixes();
-  const prefixes = prefixesData ?? NAME_PREFIX_DEFAULTS;
-  const prefix = requiredPrefix(mode, prefixes);
-
-  // Pré-preenche um complemento sugerido ao abrir (o prefixo é a badge).
-  useEffect(() => {
-    if (!open || competitionId || !competitions?.length) return;
-    const wc = findWorldCupCompetition(competitions);
-    if (wc) {
-      setCompetitionId(wc.id);
-      setName((cur) => cur || "da Copa do Mundo 2026");
-    }
-  }, [open, competitions, competitionId]);
-
-  // Federações habilitadas (teste de Confronto) não têm limite de competições.
-  const reachedLimit = !confrontoEnabled && comps.length >= MAX_COMPETITIONS_PER_LEAGUE;
-
-  async function handleAdd() {
-    if (!competitionId || !name.trim()) return;
-    try {
-      await add.mutateAsync({
-        leagueId,
-        competitionId,
-        name: `${prefix} ${name.trim()}`.trim(),
-        mode,
-        participantMode: isConfrontoNew ? participantMode : undefined,
-      });
-      toast(
-        isConfrontoNew
-          ? "Disputa criada! Agora configure e libere o sorteio."
-          : "Competição adicionada!",
-        "success",
-      );
-      setOpen(false);
-      setName("");
-      setCompetitionId("");
-    } catch (err) {
-      toast(err instanceof Error ? err.message : "Erro ao adicionar.", "error");
-    }
-  }
-
-  async function confirmDelete() {
-    if (!toDelete) return;
-    try {
-      await del.mutateAsync({ leagueId, lcId: toDelete.id });
-      toast(`Competição "${toDelete.name}" removida.`, "success");
-      setToDelete(null);
-    } catch (err) {
-      toast(err instanceof Error ? err.message : "Erro ao remover.", "error");
-    }
-  }
-
-  return (
-    <div className="space-y-3">
-      {comps.map((c) => (
-        <Card key={c.id} className="flex items-center gap-2 p-3.5">
-          <span className="min-w-0 flex-1 truncate font-semibold text-ink-900">{c.name}</span>
-          <Badge tone="neutral">
-            {c.mode === "liga"
-              ? "Liga"
-              : c.mode === "cup"
-                ? "Copa"
-                : c.mode === "table"
-                  ? "Tabela"
-                  : "Pontos"}
-          </Badge>
-          <Button
-            size="icon"
-            variant="ghost"
-            aria-label="Remover competição"
-            onClick={() => setToDelete({ id: c.id, name: c.name })}
-          >
-            <Trash2 className="size-4 text-flame-500" />
-          </Button>
-        </Card>
-      ))}
-
-      {reachedLimit ? (
-        <p className="rounded-md bg-ink-50 px-3 py-2 text-xs text-ink-500">
-          Limite inicial de {MAX_COMPETITIONS_PER_LEAGUE} competição por federação.
-          Remova a atual para trocar por outra.
-        </p>
-      ) : open ? (
-        <Card className="space-y-3 p-4">
-          <select
-            aria-label="Competição"
-            value={competitionId}
-            onChange={(e) => setCompetitionId(e.target.value)}
-            className="h-11 w-full rounded-md border border-ink-200 bg-surface px-3 outline-none focus:border-brand-500"
-          >
-            <option value="">Escolher competição…</option>
-            {competitions?.map((c) => (
-              <option key={c.id} value={c.id}>
-                {c.name}
-              </option>
-            ))}
-          </select>
-          <div className="flex items-stretch overflow-hidden rounded-md border border-ink-200 bg-surface focus-within:border-brand-500">
-            <span className="flex shrink-0 items-center bg-ink-100 px-3 text-sm font-bold text-ink-600">
-              {prefix}
-            </span>
-            <input
-              value={name}
-              onChange={(e) => setName(e.target.value)}
-              placeholder="dos amigos"
-              className="h-11 min-w-0 flex-1 bg-transparent px-3 outline-none"
-            />
-          </div>
-          <p className="-mt-1.5 text-xs text-ink-400">
-            O tipo (<span className="font-semibold text-ink-600">{prefix}</span>) já entra no nome — é
-            só completar.
-          </p>
-          {confrontoEnabled ? (
-            <div className="space-y-2">
-              <label className="text-sm font-medium text-ink-800">Tipo de disputa</label>
-              <SegmentedControl<"pontos" | "confronto">
-                value={tipo}
-                onChange={setTipo}
-                options={[
-                  { value: "pontos", label: "Pontos" },
-                  { value: "confronto", label: "Confronto" },
-                ]}
-              />
-              {tipo === "confronto" && (
-                <SegmentedControl<"liga" | "cup">
-                  value={formato}
-                  onChange={setFormato}
-                  options={[
-                    { value: "liga", label: "Liga" },
-                    { value: "cup", label: "Copa" },
-                  ]}
-                />
-              )}
-              <p className="text-xs leading-snug text-ink-500">
-                {tipo === "pontos"
-                  ? "Corrida de pontos: todo mundo acumula, quem somou mais lidera."
-                  : formato === "liga"
-                    ? "Liga: todos contra todos; cada rodada vale 3/1/0 e forma uma tabela."
-                    : "Copa: mata-mata; quem perde o confronto está fora."}
-              </p>
-              {tipo === "confronto" && (
-                <div className="space-y-1 pt-1">
-                  <label className="text-sm font-medium text-ink-800">Quem entra</label>
-                  <SegmentedControl<"admin" | "optin">
-                    value={participantMode}
-                    onChange={setParticipantMode}
-                    options={[
-                      { value: "admin", label: "Admin escolhe" },
-                      { value: "optin", label: "Cada um aceita" },
-                    ]}
-                  />
-                  <p className="text-xs leading-snug text-ink-500">
-                    {participantMode === "admin"
-                      ? "No sorteio, você marca quais membros entram."
-                      : "Cada membro confirma que quer jogar; você sorteia com quem topou."}
-                  </p>
-                </div>
-              )}
-              {tipo === "confronto" && (
-                <p className="rounded-md bg-brand-500/10 px-3 py-2 text-[11px] leading-relaxed text-brand-700">
-                  Depois de criar, você <span className="font-semibold">sorteia os confrontos</span>.
-                  Isso trava os participantes: quem entrar depois não joga esta disputa.
-                </p>
-              )}
-            </div>
-          ) : (
-            <div className="space-y-1.5">
-              <p className="text-xs font-semibold text-ink-700">Modo de disputa</p>
-              <div className="flex gap-1 rounded-pill bg-ink-100 p-1">
-                <button
-                  type="button"
-                  className="flex-1 rounded-pill bg-surface px-3 py-1.5 text-sm font-semibold text-ink-950 shadow-[var(--shadow-soft)]"
-                >
-                  Pontos
-                </button>
-                <button
-                  type="button"
-                  disabled
-                  className="flex-1 cursor-not-allowed rounded-pill px-3 py-1.5 text-xs font-semibold text-ink-400"
-                  title="Em breve"
-                >
-                  Liga · em breve
-                </button>
-                <button
-                  type="button"
-                  disabled
-                  className="flex-1 cursor-not-allowed rounded-pill px-3 py-1.5 text-xs font-semibold text-ink-400"
-                  title="Em breve"
-                >
-                  Copa · em breve
-                </button>
-              </div>
-              <p className="text-xs leading-snug text-ink-500">
-                Corrida de pontos: soma tudo numa classificação única. Os modos
-                <strong> Liga </strong> (confronto direto) e <strong>Copa</strong> (mata-mata) chegam
-                em breve.
-              </p>
-            </div>
-          )}
-          <div className="flex gap-2">
-            <Button variant="ghost" fullWidth onClick={() => setOpen(false)}>
-              Cancelar
-            </Button>
-            <Button fullWidth loading={add.isPending} onClick={handleAdd}>
-              {isConfrontoNew ? "Criar e configurar" : "Adicionar"}
-            </Button>
-          </div>
-        </Card>
-      ) : (
-        <Button variant="outline" fullWidth onClick={() => setOpen(true)}>
-          <Plus className="size-4" /> Adicionar competição
-        </Button>
-      )}
-
-      <ConfirmDialog
-        open={!!toDelete}
-        title="Remover competição"
-        message={`Remover "${toDelete?.name ?? ""}" da federação? A classificação dela e o histórico de palpites somem junto.`}
-        step2Message="Confirmação final: remover esta competição e o que está dentro?"
-        confirmLabel="Remover competição"
-        loading={del.isPending}
-        onConfirm={confirmDelete}
-        onCancel={() => setToDelete(null)}
-      />
-    </div>
-  );
-}
-
-/**
- * Editor da flâmula da federação: forma (3 flâmulas) + cores (sólido/listras/
- * grade/bola) e rotação. Salvo como `crest:flamula:...` em `leagues.logo_url`,
- * então o mesmo renderer (`Escudo`) pinta em qualquer lugar do app.
- */
-function EscudoStudio({
-  leagueId,
-  leagueName,
-  currentLogo,
-  onClose,
-}: {
-  leagueId: string;
-  leagueName: string;
-  currentLogo: string | null;
-  onClose: () => void;
-}) {
-  const update = useUpdateLeagueLogo();
-  const { toast } = useToast();
-  const [crest, setCrest] = useState<string>(currentLogo ?? "");
-
-  async function handleSave() {
-    try {
-      await update.mutateAsync({ leagueId, logoUrl: crest || null });
-      toast("Flâmula salva!", "success");
-      onClose();
-    } catch (e) {
-      toast(e instanceof Error ? e.message : "Erro ao salvar flâmula.", "error");
-    }
-  }
-
-  async function handleReset() {
-    try {
-      await update.mutateAsync({ leagueId, logoUrl: null });
-      toast("Flâmula voltou para o automático.", "info");
-      onClose();
-    } catch (e) {
-      toast(e instanceof Error ? e.message : "Erro ao resetar.", "error");
-    }
-  }
-
-  return (
-    <Card className="mb-4 space-y-4 p-4">
-      <p className="text-sm font-semibold text-ink-800">Flâmula da federação</p>
-
-      {/* Preview ao vivo */}
-      <div className="flex items-center gap-4 rounded-md bg-ink-50 p-3">
-        <Escudo src={crest || null} name={leagueName} size="xl" />
-        <div className="min-w-0">
-          <p className="truncate font-bold text-ink-900">{leagueName}</p>
-          <p className="text-xs text-ink-500">
-            Pré-visualização — a flâmula aparece na federação e na classificação.
-          </p>
-        </div>
-      </div>
-
-      <CrestEditor
-        kind="flamula"
-        name={leagueName}
-        initial={currentLogo}
-        shapes={FLAMULA_SHAPES}
-        allowBall
-        onChange={setCrest}
-      />
-
-      <div className="flex flex-wrap gap-2 pt-2">
-        <Button fullWidth loading={update.isPending} onClick={handleSave}>
-          Salvar flâmula
-        </Button>
-        <Button variant="ghost" onClick={handleReset} loading={update.isPending}>
-          Voltar ao automático
-        </Button>
-      </div>
-    </Card>
   );
 }
